TASK: [gw] Extract ORM repository as Postgres delegate
