+++ crates/gateway/src/repository/diesel_orm/mod.rs
@@ -1,5 +1,6 @@
 pub mod access_repo;
 pub mod db_conn;
+pub mod db_schema;
 pub mod role_repo;
 pub mod service_repo;
 pub mod user_repo;
