+++ crates/gateway/src/repository/postgres_db/access_repo.rs
@@ -1,76 +1,13 @@
-use diesel::prelude::*;
-use diesel::sql_types;
-use std::ops::DerefMut;
-use std::sync::{Arc, Mutex};
-use std::time::SystemTime;
-
 use crate::repository::access_repo::AccessRepository;
+use crate::repository::diesel_orm::access_repo::DieselServiceAccessRepo;
 use crate::repository::postgres_db::db_conn;
-use crate::repository::postgres_db::db_schema::service_accesses::dsl;
 use trust0_common::error::AppError;
 use trust0_common::model;
 
-/// Service access ORM model struct
-#[derive(Clone, Debug, AsChangeset, Insertable, Queryable, Selectable, PartialEq)]
-#[diesel(table_name = crate::repository::postgres_db::db_schema::service_accesses)]
-#[diesel(check_for_backend(diesel::pg::Pg))]
-pub struct ServiceAccess {
-    /// Service ID (unique across services)
-    pub service_id: i64,
-    /// RBAC entity type
-    pub entity_type: String,
-    /// Entity ID (either role ID or user ID)
-    pub entity_id: i64,
-    /// Datetime record was created
-    pub created_at: Option<SystemTime>,
-    /// Datetime record was last updated
-    pub updated_at: Option<SystemTime>,
-}
-
-fn entity_type_to_string(model_entity_type: &model::access::EntityType) -> String {
-    match model_entity_type {
-        model::access::EntityType::Role => "Role".to_string(),
-        model::access::EntityType::User => "User".to_string(),
-        _ => panic!("Invalid entity type: val={:?}", model_entity_type),
-    }
-}
-
-impl From<model::access::ServiceAccess> for ServiceAccess {
-    fn from(access: model::access::ServiceAccess) -> Self {
-        Self {
-            service_id: access.service_id,
-            entity_type: entity_type_to_string(&access.entity_type),
-            entity_id: access.entity_id,
-            created_at: None,
-            updated_at: None,
-        }
-    }
-}
-
-impl From<ServiceAccess> for model::access::ServiceAccess {
-    fn from(access: ServiceAccess) -> Self {
-        Self::from(&access)
-    }
-}
-
-impl From<&ServiceAccess> for model::access::ServiceAccess {
-    fn from(access: &ServiceAccess) -> Self {
-        Self {
-            service_id: access.service_id,
-            entity_type: match access.entity_type.as_str() {
-                "Role" => model::access::EntityType::Role,
-                "User" => model::access::EntityType::User,
-                val => panic!("Invalid ServiceAccess entity type: val={}", val),
-            },
-            entity_id: access.entity_id,
-        }
-    }
-}
-
 /// ServiceAccess Repository
 pub struct PostgresServiceAccessRepo {
-    /// If connected, an establish Postgres connection object
-    connection: Option<Arc<Mutex<PgConnection>>>,
+    /// Access repository ORM delegate
+    access_repo_delegate: Option<Box<dyn AccessRepository>>,
 }
 
 impl PostgresServiceAccessRepo {
@@ -81,18 +18,20 @@ impl PostgresServiceAccessRepo {
     /// A newly constructed [`PostgresServiceAccessRepo`] object.
     ///
     pub fn new() -> PostgresServiceAccessRepo {
-        PostgresServiceAccessRepo { connection: None }
+        PostgresServiceAccessRepo {
+            access_repo_delegate: None,
+        }
     }
 }
 
 impl AccessRepository for PostgresServiceAccessRepo {
     fn connect_to_datasource(&mut self, connect_spec: &str) -> Result<(), AppError> {
-        self.connection = Some(
-            db_conn::INSTANCE
+        self.access_repo_delegate = Some(Box::new(DieselServiceAccessRepo::new(
+            &db_conn::INSTANCE
                 .lock()
                 .unwrap()
                 .establish_connection(connect_spec)?,
-        );
+        )));
         Ok(())
     }
 
@@ -100,32 +39,7 @@ impl AccessRepository for PostgresServiceAccessRepo {
         &self,
         access: model::access::ServiceAccess,
     ) -> Result<model::access::ServiceAccess, AppError> {
-        let access_entity: ServiceAccess = access.clone().into();
-
-        // Delete (if necess)
-        _ = self.delete(access.service_id, &access.entity_type, access.entity_id)?;
-
-        // Insert
-        diesel::insert_into(dsl::service_accesses)
-            .values((
-                dsl::service_id.eq(access_entity.service_id),
-                dsl::entity_type.eq(access_entity.entity_type.as_str()),
-                dsl::entity_id.eq(access_entity.entity_id),
-            ))
-            .returning(ServiceAccess::as_returning())
-            .get_result::<ServiceAccess>(
-                self.connection
-                    .as_ref()
-                    .unwrap()
-                    .lock()
-                    .unwrap()
-                    .deref_mut(),
-            )
-            .map_err(|err| {
-                AppError::GenWithMsgAndErr("Error putting ServiceAccess".to_string(), Box::new(err))
-            })?;
-
-        Ok(access)
+        self.access_repo_delegate.as_ref().unwrap().put(access)
     }
 
     fn get(
@@ -134,31 +48,10 @@ impl AccessRepository for PostgresServiceAccessRepo {
         entity_type: &model::access::EntityType,
         entity_id: i64,
     ) -> Result<Option<model::access::ServiceAccess>, AppError> {
-        let entity_type = entity_type_to_string(entity_type);
-        match dsl::service_accesses
-            .find((entity_type.as_str(), entity_id, service_id))
-            .select(ServiceAccess::as_select())
-            .first(
-                self.connection
-                    .as_ref()
-                    .unwrap()
-                    .lock()
-                    .unwrap()
-                    .deref_mut(),
-            ) {
-            Ok(access) => {
-                let access: model::access::ServiceAccess = access.into();
-                Ok(Some(access))
-            }
-            Err(diesel::NotFound) => Ok(None),
-            Err(err) => Err(AppError::GenWithMsgAndErr(
-                format!(
-                    "Error getting ServiceAccess: svc_id={}, type={:?}, ent_id={}",
-                    service_id, entity_type, entity_id
-                ),
-                Box::new(err),
-            )),
-        }
+        self.access_repo_delegate
+            .as_ref()
+            .unwrap()
+            .get(service_id, entity_type, entity_id)
     }
 
     fn get_for_user(
@@ -166,77 +59,20 @@ impl AccessRepository for PostgresServiceAccessRepo {
         service_id: i64,
         user: &model::user::User,
     ) -> Result<Option<model::access::ServiceAccess>, AppError> {
-        let f = false.into_sql::<sql_types::Bool>();
-        let access_list: Vec<ServiceAccess> = dsl::service_accesses
-            .filter(
-                dsl::service_id.eq(service_id).and(
-                    f.or(dsl::entity_type
-                        .eq("User")
-                        .and(dsl::entity_id.eq(user.user_id)))
-                        .or(dsl::entity_type
-                            .eq("Role")
-                            .and(dsl::entity_id.eq_any(user.roles.as_slice()))),
-                ),
-            )
-            .select(ServiceAccess::as_select())
-            .load(
-                self.connection
-                    .as_ref()
-                    .unwrap()
-                    .lock()
-                    .unwrap()
-                    .deref_mut(),
-            )
-            .map_err(|err| {
-                AppError::GenWithMsgAndErr(
-                    format!(
-                        "Error getting ServiceAccess: svc_id={}, user_id={}",
-                        service_id, user.user_id
-                    ),
-                    Box::new(err),
-                )
-            })?;
-
-        Ok(access_list.first().map(|access| {
-            let access: model::access::ServiceAccess = access.into();
-            access
-        }))
+        self.access_repo_delegate
+            .as_ref()
+            .unwrap()
+            .get_for_user(service_id, user)
     }
 
     fn get_all_for_user(
         &self,
         user: &model::user::User,
     ) -> Result<Vec<model::access::ServiceAccess>, AppError> {
-        let f = false.into_sql::<sql_types::Bool>();
-        let access_list: Vec<ServiceAccess> = dsl::service_accesses
-            .filter(
-                f.or(dsl::entity_type
-                    .eq("User")
-                    .and(dsl::entity_id.eq(user.user_id)))
-                    .or(dsl::entity_type
-                        .eq("Role")
-                        .and(dsl::entity_id.eq_any(user.roles.as_slice()))),
-            )
-            .select(ServiceAccess::as_select())
-            .load(
-                self.connection
-                    .as_ref()
-                    .unwrap()
-                    .lock()
-                    .unwrap()
-                    .deref_mut(),
-            )
-            .map_err(|err| {
-                AppError::GenWithMsgAndErr(
-                    format!("Error getting ServiceAccess: user_id={}", user.user_id),
-                    Box::new(err),
-                )
-            })?;
-
-        Ok(access_list
-            .iter()
-            .map(|access| access.into())
-            .collect::<Vec<model::access::ServiceAccess>>())
+        self.access_repo_delegate
+            .as_ref()
+            .unwrap()
+            .get_all_for_user(user)
     }
 
     fn delete(
@@ -245,38 +81,10 @@ impl AccessRepository for PostgresServiceAccessRepo {
         entity_type: &model::access::EntityType,
         entity_id: i64,
     ) -> Result<Option<model::access::ServiceAccess>, AppError> {
-        let curr_access = self.get(service_id, entity_type, entity_id)?;
-        if curr_access.is_none() {
-            return Ok(None);
-        }
-
-        let entity_type = entity_type_to_string(entity_type);
-
-        match diesel::delete(dsl::service_accesses)
-            .filter(
-                dsl::entity_type
-                    .eq(entity_type.as_str())
-                    .and(dsl::entity_id.eq(entity_id))
-                    .and(dsl::service_id.eq(service_id)),
-            )
-            .execute(
-                self.connection
-                    .as_ref()
-                    .unwrap()
-                    .lock()
-                    .unwrap()
-                    .deref_mut(),
-            ) {
-            Ok(_) => Ok(curr_access),
-            Err(diesel::NotFound) => Ok(None),
-            Err(err) => Err(AppError::GenWithMsgAndErr(
-                format!(
-                    "Error getting ServiceAccess: svc_id={}, type={:?}, ent_id={}",
-                    service_id, entity_type, entity_id
-                ),
-                Box::new(err),
-            )),
-        }
+        self.access_repo_delegate
+            .as_ref()
+            .unwrap()
+            .delete(service_id, entity_type, entity_id)
     }
 }
 
@@ -284,814 +92,158 @@ impl AccessRepository for PostgresServiceAccessRepo {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use serial_test::serial;
-    use std::collections::HashMap;
-    use std::path::PathBuf;
-
-    pub const POSTGRES_DATABASE_DIR_PATHPARTS: [&str; 8] = [
-        env!("CARGO_MANIFEST_DIR"),
-        "..",
-        "..",
-        "target",
-        "test-gateway",
-        "postgres",
-        "data",
-        "access",
-    ];
+    use crate::repository::access_repo::tests::MockAccessRepo;
+    use mockall::predicate;
 
     #[test]
-    fn pgdbaccessrepo_access_from_model() {
-        let model_access1 = model::access::ServiceAccess {
-            service_id: 200,
-            entity_type: model::access::EntityType::Role,
-            entity_id: 50,
-        };
-        let model_access2 = model::access::ServiceAccess {
-            service_id: 201,
-            entity_type: model::access::EntityType::User,
-            entity_id: 100,
-        };
-        let expected_access1 = ServiceAccess {
-            service_id: 200,
-            entity_type: "Role".to_string(),
-            entity_id: 50,
-            created_at: None,
-            updated_at: None,
-        };
-        let expected_access2 = ServiceAccess {
-            service_id: 201,
-            entity_type: "User".to_string(),
-            entity_id: 100,
-            created_at: None,
-            updated_at: None,
-        };
-        assert_eq!(ServiceAccess::from(model_access1), expected_access1);
-        assert_eq!(ServiceAccess::from(model_access2), expected_access2);
+    fn pgdbaccessrepo_connect_to_datasource() {
+        let mut access_repo = PostgresServiceAccessRepo::new();
+        if let Ok(()) = access_repo.connect_to_datasource("INVALID") {
+            panic!("Unexpected successful result");
+        }
     }
 
     #[test]
-    fn pgdbaccessrepo_access_to_model() {
-        let access1 = ServiceAccess {
-            service_id: 200,
-            entity_type: "Role".to_string(),
-            entity_id: 50,
-            created_at: None,
-            updated_at: None,
-        };
-        let access2 = ServiceAccess {
-            service_id: 201,
-            entity_type: "User".to_string(),
-            entity_id: 100,
-            created_at: None,
-            updated_at: None,
-        };
-        let expected_model_access1 = model::access::ServiceAccess {
-            service_id: 200,
-            entity_type: model::access::EntityType::Role,
-            entity_id: 50,
-        };
-        let expected_model_access2 = model::access::ServiceAccess {
-            service_id: 201,
-            entity_type: model::access::EntityType::User,
-            entity_id: 100,
-        };
-        assert_eq!(
-            model::access::ServiceAccess::from(access1),
-            expected_model_access1
-        );
-        assert_eq!(
-            model::access::ServiceAccess::from(access2),
-            expected_model_access2
-        );
-    }
-
-    #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_put_when_existing_access() {
+    fn pgdbaccessrepo_put() {
         let expected_access = model::access::ServiceAccess {
             service_id: 200,
             entity_type: model::access::EntityType::User,
             entity_id: 100,
         };
-
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
-
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
-
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
-
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
-
-            let result = access_repo.put(expected_access.clone());
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-
-            let access = result.unwrap();
-
-            assert_eq!(access, expected_access);
-        }
-    }
-
-    #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_put_when_new_access() {
-        let expected_access = model::access::ServiceAccess {
-            service_id: 201,
-            entity_type: model::access::EntityType::Role,
-            entity_id: 51,
+        let expected_access_copy = expected_access.clone();
+        let mut access_repo_delegate = MockAccessRepo::new();
+        access_repo_delegate
+            .expect_put()
+            .with(predicate::eq(expected_access.clone()))
+            .times(1)
+            .return_once(|_| Ok(expected_access_copy));
+        let access_repo = PostgresServiceAccessRepo {
+            access_repo_delegate: Some(Box::new(access_repo_delegate)),
         };
 
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
-
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
-
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
-
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
-
-            let result = access_repo.put(expected_access.clone());
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-
-            let access = result.unwrap();
-            assert_eq!(access, expected_access);
+        if let Err(err) = access_repo.put(expected_access) {
+            panic!("Unexpected result: err={:?}", &err);
         }
     }
 
     #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_get_when_invalid_user() {
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
-
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
-
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
-
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
-
-            let result = access_repo.get(200, &model::access::EntityType::User, 10);
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-
-            assert!(result.unwrap().is_none());
-        }
-    }
-
-    #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_get_when_invalid_service() {
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
-
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
-
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
-
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
-
-            let result = access_repo.get(20, &model::access::EntityType::Role, 50);
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-
-            assert!(result.unwrap().is_none());
-        }
-    }
-
-    #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_get_when_valid_user_and_service() {
+    fn pgdbaccessrepo_get() {
         let expected_access = model::access::ServiceAccess {
             service_id: 200,
             entity_type: model::access::EntityType::User,
             entity_id: 100,
         };
+        let expected_access_copy = expected_access.clone();
+        let mut access_repo_delegate = MockAccessRepo::new();
+        access_repo_delegate
+            .expect_get()
+            .with(
+                predicate::eq(expected_access.service_id),
+                predicate::eq(expected_access.entity_type.clone()),
+                predicate::eq(expected_access.entity_id),
+            )
+            .times(1)
+            .return_once(|_, _, _| Ok(Some(expected_access_copy)));
+        let access_repo = PostgresServiceAccessRepo {
+            access_repo_delegate: Some(Box::new(access_repo_delegate)),
+        };
 
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
-
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
-
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
-
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
-
-            let result = access_repo.get(200, &model::access::EntityType::User, 100);
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-
-            let actual_access = result.unwrap();
-
-            assert!(actual_access.is_some());
-            assert_eq!(actual_access.unwrap(), expected_access);
+        if let Err(err) = access_repo.get(
+            expected_access.service_id,
+            &expected_access.entity_type.clone(),
+            expected_access.entity_id,
+        ) {
+            panic!("Unexpected result: err={:?}", &err);
         }
     }
 
     #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_get_all_for_user_when_invalid_user() {
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
-
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
-
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
-
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
-
-            let result = access_repo.get_all_for_user(&model::user::User::new(
-                10,
-                Some("uname10"),
-                Some("pass10"),
-                "name10",
-                &model::user::Status::Active,
-                &vec![],
-            ));
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-
-            assert_eq!(result.unwrap().len(), 0);
-        }
-    }
-
-    #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_get_all_for_user_when_valid_user() {
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
-
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
-
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
-
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
-
-            let result = access_repo.get_all_for_user(&model::user::User::new(
-                100,
-                Some("uname1"),
-                Some("pass1"),
-                "name1",
-                &model::user::Status::Active,
-                &vec![50],
-            ));
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-
-            let actual_accesses = result.unwrap();
-            assert_eq!(actual_accesses.len(), 2);
-
-            let expected_access_db_map: HashMap<
-                (i64, model::access::EntityType, i64),
-                model::access::ServiceAccess,
-            > = HashMap::from([
-                (
-                    (200, model::access::EntityType::User, 100),
-                    model::access::ServiceAccess {
-                        service_id: 200,
-                        entity_type: model::access::EntityType::User,
-                        entity_id: 100,
-                    },
-                ),
-                (
-                    (201, model::access::EntityType::Role, 50),
-                    model::access::ServiceAccess {
-                        service_id: 200,
-                        entity_type: model::access::EntityType::User,
-                        entity_id: 100,
-                    },
-                ),
-            ]);
-
-            assert_eq!(
-                actual_accesses
-                    .iter()
-                    .filter(|entry| !expected_access_db_map.contains_key(&(
-                        entry.service_id,
-                        entry.entity_type.clone(),
-                        entry.entity_id
-                    )))
-                    .count(),
-                0
-            );
-        }
-    }
-
-    #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_get_for_user_when_invalid_user() {
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
-
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
-
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
-
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
-
-            let result = access_repo.get_for_user(
-                200,
-                &model::user::User::new(
-                    10,
-                    Some("uname10"),
-                    Some("pass10"),
-                    "name10",
-                    &model::user::Status::Active,
-                    &vec![5],
-                ),
-            );
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-
-            assert!(result.unwrap().is_none());
-        }
-    }
-
-    #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_get_for_user_when_valid_user() {
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
-
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
-
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
-
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
-
-            let result = access_repo.get_for_user(
-                200,
-                &model::user::User::new(
-                    100,
-                    Some("uname10"),
-                    Some("pass10"),
-                    "name10",
-                    &model::user::Status::Active,
-                    &vec![5],
-                ),
-            );
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-            let actual_access = result.unwrap();
-
-            assert!(actual_access.is_some());
-
-            let expected_access = model::access::ServiceAccess {
-                service_id: 200,
-                entity_type: model::access::EntityType::User,
-                entity_id: 100,
-            };
-
-            assert_eq!(actual_access.unwrap(), expected_access);
-        }
-    }
-
-    #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_get_for_user_when_valid_role() {
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
-
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
-
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
-
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
-
-            let result = access_repo.get_for_user(
-                201,
-                &model::user::User::new(
-                    10,
-                    Some("uname10"),
-                    Some("pass10"),
-                    "name10",
-                    &model::user::Status::Active,
-                    &vec![50],
-                ),
-            );
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-            let actual_access = result.unwrap();
-
-            assert!(actual_access.is_some());
-
-            let expected_access = model::access::ServiceAccess {
-                service_id: 201,
-                entity_type: model::access::EntityType::Role,
-                entity_id: 50,
-            };
-
-            assert_eq!(actual_access.unwrap(), expected_access);
-        }
-    }
-
-    #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_delete_when_invalid_access() {
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
-
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
-
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
-
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
-
-            let result = access_repo.delete(200, &model::access::EntityType::User, 10);
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-
-            assert!(result.unwrap().is_none());
-        }
-    }
-
-    #[test]
-    #[serial(pgdb_access)]
-    fn pgdbaccessrepo_delete_when_valid_access() {
-        let expected_prev_access = model::access::ServiceAccess {
+    fn pgdbaccessrepo_get_for_user() {
+        let expected_access = model::access::ServiceAccess {
             service_id: 200,
             entity_type: model::access::EntityType::User,
             entity_id: 100,
         };
+        let expected_access_copy = expected_access.clone();
+        let expected_user = model::user::User {
+            user_id: 100,
+            name: "User100".to_string(),
+            status: model::user::Status::Active,
+            roles: vec![50, 51],
+            user_name: Some("uname100".to_string()),
+            password: Some("pass100".to_string()),
+        };
+        let mut access_repo_delegate = MockAccessRepo::new();
+        access_repo_delegate
+            .expect_get_for_user()
+            .with(
+                predicate::eq(expected_access.service_id),
+                predicate::eq(expected_user.clone()),
+            )
+            .times(1)
+            .return_once(|_, _| Ok(Some(expected_access_copy)));
+        let access_repo = PostgresServiceAccessRepo {
+            access_repo_delegate: Some(Box::new(access_repo_delegate)),
+        };
 
-        let database_dir: PathBuf = POSTGRES_DATABASE_DIR_PATHPARTS.iter().collect();
-        let (pg_embed, mut db_conn) = db_conn::tests::DB
-            .lock()
-            .unwrap()
-            .setup_db(database_dir.clone())
-            .unwrap();
+        if let Err(err) = access_repo.get_for_user(expected_access.service_id, &expected_user) {
+            panic!("Unexpected result: err={:?}", &err);
+        }
+    }
 
-        {
-            let _transaction = db_conn::tests::EmbeddedDbTransaction {
-                database_dir,
-                pg_embed: pg_embed.clone(),
-            };
+    #[test]
+    fn pgdbaccessrepo_get_all_for_user() {
+        let expected_user = model::user::User {
+            user_id: 100,
+            name: "User100".to_string(),
+            status: model::user::Status::Active,
+            roles: vec![50, 51],
+            user_name: Some("uname100".to_string()),
+            password: Some("pass100".to_string()),
+        };
+        let mut access_repo_delegate = MockAccessRepo::new();
+        access_repo_delegate
+            .expect_get_all_for_user()
+            .with(predicate::eq(expected_user.clone()))
+            .times(1)
+            .return_once(|_| Ok(Vec::new()));
+        let access_repo = PostgresServiceAccessRepo {
+            access_repo_delegate: Some(Box::new(access_repo_delegate)),
+        };
 
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(&mut db_conn, &db_conn::tests::SQL_CREATE_SERVICE_RECORDS)
-                .unwrap();
-            db_conn::tests::DB
-                .lock()
-                .unwrap()
-                .execute_sql(
-                    &mut db_conn,
-                    &db_conn::tests::SQL_CREATE_SERVICE_ACCESS_RECORDS,
-                )
-                .unwrap();
+        if let Err(err) = access_repo.get_all_for_user(&expected_user) {
+            panic!("Unexpected result: err={:?}", &err);
+        }
+    }
 
-            let mut access_repo = PostgresServiceAccessRepo::new();
-            access_repo
-                .connect_to_datasource(
-                    &pg_embed
-                        .lock()
-                        .unwrap()
-                        .full_db_uri(db_conn::tests::DB_NAME),
-                )
-                .unwrap();
+    #[test]
+    fn pgdbaccessrepo_delete() {
+        let expected_access = model::access::ServiceAccess {
+            service_id: 200,
+            entity_type: model::access::EntityType::User,
+            entity_id: 100,
+        };
+        let expected_access_copy = expected_access.clone();
+        let mut access_repo_delegate = MockAccessRepo::new();
+        access_repo_delegate
+            .expect_delete()
+            .with(
+                predicate::eq(expected_access.service_id),
+                predicate::eq(expected_access.entity_type.clone()),
+                predicate::eq(expected_access.entity_id),
+            )
+            .times(1)
+            .return_once(|_, _, _| Ok(Some(expected_access_copy)));
+        let access_repo = PostgresServiceAccessRepo {
+            access_repo_delegate: Some(Box::new(access_repo_delegate)),
+        };
 
-            let result = access_repo.delete(200, &model::access::EntityType::User, 100);
-
-            if let Err(err) = &result {
-                panic!("Unexpected result: err={:?}", &err)
-            }
-
-            let actual_prev_access = result.unwrap();
-
-            assert!(actual_prev_access.is_some());
-            assert_eq!(actual_prev_access.unwrap(), expected_prev_access);
+        if let Err(err) = access_repo.delete(
+            expected_access.service_id,
+            &expected_access.entity_type,
+            expected_access.entity_id,
+        ) {
+            panic!("Unexpected result: err={:?}", &err);
         }
     }
 }
